ValuePlugTest : Adopt relevant tests from CompoundPlugTest.
CompoundPlug will be removed in the future, but most of its implementation is in ValuePlug, and we'll want to keep the tests when we remove CompoundPlug (and CompoundPlugTest).
